+++ lib/commands/moveStalledJobsToWait-7.lua
@@ -22,17 +22,32 @@
 
 local rcall = redis.call
 
-local function batches(n, batchSize)
-  local i = 0
+-- Includes
+--- @include "includes/batches"
+--- @include "includes/getTargetQueueList"
+--- @include "includes/removeDebounceKeyIfNeeded"
 
-  return function()
-    local from = i * batchSize + 1
-    i = i + 1
-    if (from <= n) then
-      local to = math.min(from + batchSize - 1, n)
-      return from, to
-    end
+local function removeJob(jobId, baseKey)
+  local jobKey = baseKey .. jobId
+  rcall("DEL", jobKey, jobKey .. ':logs')
+end
+
+local function removeJobsByMaxAge(timestamp, maxAge, targetSet, prefix)
+  local start = timestamp - maxAge * 1000
+  local jobIds = rcall("ZREVRANGEBYSCORE", targetSet, start, "-inf")
+  for i, jobId in ipairs(jobIds) do
+    removeJob(jobId, prefix)
   end
+  rcall("ZREMRANGEBYSCORE", targetSet, "-inf", start)
+end
+
+local function removeJobsByMaxCount(maxCount, targetSet, prefix)
+  local start = maxCount
+  local jobIds = rcall("ZREVRANGE", targetSet, start, -1)
+  for i, jobId in ipairs(jobIds) do
+    removeJob(jobId, prefix)
+  end
+  rcall("ZREMRANGEBYRANK", targetSet, 0, -(maxCount + 1))
 end
 
 -- Check if we need to check for stalled jobs now.
@@ -47,15 +62,6 @@ local stalling = rcall('SMEMBERS', KEYS[1])
 local stalled = {}
 local failed = {}
 if(#stalling > 0) then
-
-  local dst
-  -- wait or paused destination
-  if rcall("EXISTS", KEYS[6]) ~= 1 then
-    dst = KEYS[2]
-  else
-    dst = KEYS[7]
-  end
-
   rcall('DEL', KEYS[1])
 
   local MAX_STALLED_JOB_COUNT = tonumber(ARGV[1])
@@ -73,13 +79,44 @@ if(#stalling > 0) then
         -- If this job has been stalled too many times, such as if it crashes the worker, then fail it.
         local stalledCount = rcall("HINCRBY", jobKey, "stalledCounter", 1)
         if(stalledCount > MAX_STALLED_JOB_COUNT) then
+          local jobAttributes = rcall("HMGET", jobKey, "opts", "deid")
+          local opts = cjson.decode(jobAttributes[1])
+          local removeOnFailType = type(opts["removeOnFail"])
           rcall("ZADD", KEYS[4], ARGV[3], jobId)
-          rcall("HSET", jobKey, "failedReason", "job stalled more than allowable limit")
-          rcall("PUBLISH", KEYS[4],  "{\"jobId\":\"" .. jobId .. "\", \"val\": \"job stalled more than maxStalledCount\"}")
+          rcall("HMSET", jobKey, "failedReason", "job stalled more than allowable limit",
+            "finishedOn", ARGV[3])
+          removeDebounceKeyIfNeeded(ARGV[2], jobAttributes[2])
+          rcall("PUBLISH", KEYS[4],  '{"jobId":"' .. jobId .. '", "val": "job stalled more than maxStalledCount"}')
+
+          if removeOnFailType == "number" then
+            removeJobsByMaxCount(opts["removeOnFail"],
+                KEYS[4], ARGV[2])
+          elseif removeOnFailType == "boolean" then
+            if opts["removeOnFail"] then
+              removeJob(jobId, ARGV[2])
+              rcall("ZREM", KEYS[4], jobId)
+            end
+          elseif removeOnFailType ~= "nil" then
+            local maxAge = opts["removeOnFail"]["age"]
+            local maxCount = opts["removeOnFail"]["count"]
+
+            if maxAge ~= nil then
+              removeJobsByMaxAge(ARGV[3], maxAge,
+                KEYS[4], ARGV[2])
+            end
+
+            if maxCount ~= nil and maxCount > 0 then
+              removeJobsByMaxCount(maxCount, KEYS[4],
+                ARGV[2])
+            end
+          end
+
           table.insert(failed, jobId)
         else
+          local target = getTargetQueueList(KEYS[6], KEYS[2], KEYS[7])
+
           -- Move the job back to the wait queue, to immediately be picked up by a waiting worker.
-          rcall("RPUSH", dst, jobId)
+          rcall("RPUSH", target, jobId)
           rcall('PUBLISH', KEYS[1] .. '@', jobId)
           table.insert(stalled, jobId)
         end